+++ frontend/app_student/lib/components/header_logo.dart
@@ -6,13 +6,16 @@ class HeaderLogo extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      height: 210.0, // Hauteur du Container
+      height: 300.0, // Hauteur du Container
       color: const Color(0xFF005067),
       child: Center(
-        child: SizedBox(
-          width: 200.0,
-          height: 200.0,
-          child: Image.asset('assets/images/3il-logo.jpg'),
+        child: Padding(
+          padding: const EdgeInsets.only(top: 50.0),
+          child: SizedBox(
+            width: 200.0,
+            height: 200.0,
+            child: Image.asset('assets/images/3il-logo.jpg'),
+          ),
         ),
       ),
     );
